refactor query history by moving components to their own files
